Add obo score draft computation
