+++ util/dashboard_config.py
@@ -1,17 +1,18 @@
 #!/usr/bin/env python3
 
 import os
-import sys
 import yaml
 import click
 import logging
 import urllib.request
-from datetime import datetime
 
-from lib import DashboardConfig, runcmd, runcmd, sha256sum, save_yaml, load_yaml, robot_prepare_ontology
+
+from lib import DashboardConfig, runcmd, runcmd, sha256sum, save_yaml, \
+    load_yaml, robot_prepare_ontology, get_hours_since, get_base_prefixes, \
+    compute_percentage_reused_entities, compute_dashboard_score, round_float, \
+    compute_obo_score
 
 logging.basicConfig(level=logging.INFO)
-
 
 @click.group()
 def cli():
@@ -76,42 +77,8 @@ def rundashboard(configfile, clean):
     logging.info("Postprocess files for github")
     runcmd(f"make truncate_reports_for_github {make_parameters} -B")
 
-def get_hours_since(timestamp):
-    modified_date = datetime.fromtimestamp(timestamp)
-    now = datetime.now()
-    duration = now - modified_date
-    hours_since = (duration.total_seconds() // 3600)
-    return hours_since
-
-
-def get_prefix_from_url_namespace(ns, curie_map):
-    for prefix in curie_map:
-        if ns==curie_map[prefix]:
-            return prefix
-    raise Exception(f"Namespace {ns} not found in curie map, aborting..")
-
-def get_base_prefixes(curie_map, base_namespaces):
-    internal_ns = []
-    for ns in base_namespaces:
-        prefix = get_prefix_from_url_namespace(ns, curie_map)
-        internal_ns.append(prefix)
-    return internal_ns
-
-def compute_percentage_reused_entities(entity_use_map, internal_ns):
-    internal_entities = 0
-    external_entities = 0
-    for prefix in entity_use_map:
-        if prefix in internal_ns:
-            internal_entities += entity_use_map[prefix]
-        else:
-            external_entities += entity_use_map[prefix]
-
-    reuse_score = 100*(external_entities/(external_entities+internal_entities))
-    score_string = "%.2f" % round(reuse_score, 2)
-    return score_string+" %"
-
-
 def prepare_ontologies(ontologies, ontology_dir, dashboard_dir, make_parameters, config):
+    ontology_use = {}
     for o in ontologies:
         logging.info(f"Preparing {o}...")
         ont_path = os.path.join(ontology_dir, f"{o}-raw.owl")
@@ -233,29 +200,40 @@ def prepare_ontologies(ontologies, ontology_dir, dashboard_dir, make_parameters,
         else:
             logging.info(f"{o} has not changed since last run, skipping process.")
 
-        try:
-            metrics = load_yaml(ont_metrics_path)
-            base_prefixes = get_base_prefixes(metrics['metrics']['curie_map'], base_namespaces)
-            ont_results['base_prefixes'] = base_prefixes
-            ont_results['metrics'] = {}
-            ont_results['metrics']['Info: Logical consistency'] = metrics['metrics']['consistent']
-            ont_results['metrics']['Entities: Number of unsatisfiable classes'] = metrics['metrics']['unsatisfiable_class_count']
-            ont_results['metrics']['Axioms: Number of axioms'] = metrics['metrics']['axiom_count_incl']
-            ont_results['metrics']['Entities: Number of classes'] = metrics['metrics']['class_count_incl']
-            ont_results['metrics']['Entities: Number of object properties'] = metrics['metrics']['obj_property_count_incl']
-            ont_results['metrics']['Entities: % of entities reused'] = compute_percentage_reused_entities(metrics['metrics']['namespace_entity_count_incl'], base_prefixes)
-            ont_results['metrics']['Info: Usage of namespaces in axioms'] = metrics['metrics']['namespace_axiom_count_incl']
-            ont_results['metrics']['Entities: Number of individuals'] = metrics['metrics']['individual_count_incl']
-            ont_results['metrics']['Entities: Number of data properties'] = metrics['metrics']['dataproperty_count_incl']
-            ont_results['metrics']['Entities: Number of annotation properties'] = metrics['metrics']['annotation_property_count_incl']
-            ont_results['metrics']['Axioms: Breakdown of axiom types'] = metrics['metrics']['axiom_type_count_incl']
-            ont_results['metrics']['Info: Breakdown of OWL class expressions used'] = metrics['metrics']['class_expression_count_incl']
-            ont_results['metrics']['Info: Does the ontology fall under OWL 2 DL?'] = metrics['metrics']['owl2_dl']
-            ont_results['metrics']['Info: Syntax'] = metrics['metrics']['syntax']
-            save_yaml(ont_results, ont_results_path)
-        except Exception:
-            logging.exception(f'Missing or broken metrics file for {o}: {ont_metrics_path}')
-            ont_results['failure'] = 'broken_metrics_file'
+        if os.path.exists(ont_metrics_path):
+            try:
+                metrics = load_yaml(ont_metrics_path)
+                base_prefixes = get_base_prefixes(metrics['metrics']['curie_map'], base_namespaces)
+                ont_results['base_prefixes'] = base_prefixes
+                ont_results['metrics'] = {}
+                ont_results['metrics']['Info: Logical consistency'] = metrics['metrics']['consistent']
+                ont_results['metrics']['Entities: Number of unsatisfiable classes'] = metrics['metrics']['unsatisfiable_class_count']
+                ont_results['metrics']['Axioms: Number of axioms'] = metrics['metrics']['axiom_count_incl']
+                ont_results['metrics']['Entities: Number of classes'] = metrics['metrics']['class_count_incl']
+                ont_results['metrics']['Entities: Number of object properties'] = metrics['metrics']['obj_property_count_incl']
+                ont_results['metrics']['Entities: % of entities reused'] = compute_percentage_reused_entities(metrics['metrics']['namespace_entity_count_incl'], base_prefixes)
+                ont_results['metrics']['Info: Usage of namespaces in axioms'] = metrics['metrics']['namespace_axiom_count_incl']
+                ont_results['metrics']['Entities: Number of individuals'] = metrics['metrics']['individual_count_incl']
+                ont_results['metrics']['Entities: Number of data properties'] = metrics['metrics']['dataproperty_count_incl']
+                ont_results['metrics']['Entities: Number of annotation properties'] = metrics['metrics']['annotation_property_count_incl']
+                ont_results['metrics']['Axioms: Breakdown of axiom types'] = metrics['metrics']['axiom_type_count_incl']
+                ont_results['metrics']['Info: Breakdown of OWL class expressions used'] = metrics['metrics']['class_expression_count_incl']
+                ont_results['metrics']['Info: Does the ontology fall under OWL 2 DL?'] = metrics['metrics']['owl2_dl']
+                ont_results['metrics']['Info: Syntax'] = metrics['metrics']['syntax']
+                for base_prefix in base_prefixes:
+                    for used_prefix in ont_results['metrics']['Info: Usage of namespaces in axioms']:
+                        if used_prefix not in ontology_use:
+                            ontology_use[used_prefix] = []
+                        ontology_use[used_prefix].append(base_prefix)
+                save_yaml(ont_results, ont_results_path)
+            except Exception:
+                logging.exception(f'Broken metrics file for {o}: {ont_metrics_path}')
+                ont_results['failure'] = 'broken_metrics_file'
+                save_yaml(ont_results, ont_results_path)
+                continue
+        else:
+            logging.exception(f'Missing metrics file for {o}: {ont_metrics_path}')
+            ont_results['failure'] = 'missing_metrics_file'
             save_yaml(ont_results, ont_results_path)
             continue
 
@@ -270,6 +248,34 @@ def prepare_ontologies(ontologies, ontology_dir, dashboard_dir, make_parameters,
             ont_results['failure'] = 'inconsistent_ontology'
             save_yaml(ont_results, ont_results_path)
             continue
+
+    # This has to be done after all ontologies are analysed, because we need their usage information to quantify impact.
+    oboscore_weights = config.get_oboscore_weights()
+    oboscore_maximpacts = config.get_oboscore_max_impact()
+
+    for o in ontologies:
+        ont_dashboard_dir = os.path.join(dashboard_dir, o)
+        ont_results_path = os.path.join(ont_dashboard_dir, "dashboard.yml")
+        with open(ont_results_path, 'r') as f:
+            ont_results = yaml.load(f, Loader=yaml.SafeLoader)
+        if 'metrics' in ont_results:
+            uses = []
+            if 'base_prefixes' in ont_results:
+                for base_prefix in ont_results['base_prefixes']:
+                    if base_prefix in ontology_use:
+                        uses.extend(ontology_use[base_prefix])
+            uses = list(set(uses))
+
+            ont_results['metrics']['Info: How many ontologies use it?'] = len(uses)
+            dashboard_score = {}
+            dashboard_score['_impact'] = round_float(float(ont_results['metrics']['Info: How many ontologies use it?'])/len(ontologies))
+            dashboard_score['_reuse'] = round_float(float(ont_results['metrics']['Entities: % of entities reused'])/100)
+            dashboard_score['_dashboard'] = round_float(float(compute_dashboard_score(ont_results, oboscore_weights, oboscore_maximpacts))/100)
+            oboscore = compute_obo_score(dashboard_score['_impact'], dashboard_score['_reuse'], dashboard_score['_dashboard'], oboscore_weights)
+            dashboard_score['oboscore'] = round_float(oboscore['score'])
+            dashboard_score['_formula'] = oboscore['formula']
+            ont_results['metrics']['Info: Experimental OBO score'] = dashboard_score
+            save_yaml(ont_results, ont_results_path)
 
         dashboard_html = os.path.join(ont_dashboard_dir, "dashboard.html")
