+++ util/lib.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import os
 import yaml
 import logging
 import subprocess
@@ -8,7 +7,7 @@ import urllib.request
 import hashlib
 from subprocess import check_call
 import requests
-from pathlib import Path
+from datetime import datetime
 
 def runcmd(cmd):
     logging.info("RUNNING: {}".format(cmd))
@@ -55,10 +54,13 @@ class DashboardConfig:
         weights['report_errors'] = 0.05
         weights['report_warning'] = 0.01
         weights['report_info'] = 0.005
+        weights['dashboard'] = 2
+        weights['impact'] = 3
+        weights['reuse'] = 1
         if 'obo_score_weights' in self.config:
             for weight in self.config['obo_score_weights']:
-                if 'impact_factor' in self.config['obo_score_weights']:
-                    weights[weight] = self.config['obo_score_weights']['impact_factor'][weight]
+                if 'impact_factor' in self.config['obo_score_weights'][weight]:
+                    weights[weight] = self.config['obo_score_weights'][weight]['impact_factor']
         return weights
 
     def get_oboscore_max_impact(self):
@@ -70,10 +72,13 @@ class DashboardConfig:
         weights['report_errors'] = 10
         weights['report_warning'] = 5
         weights['report_info'] = 2
+        weights['dashboard'] = 2
+        weights['impact'] = 3
+        weights['reuse'] = 1
         if 'obo_score_weights' in self.config:
             for weight in self.config['obo_score_weights']:
-                if 'max_impact' in self.config['obo_score_weights']:
-                    weights[weight] = self.config['obo_score_weights']['max_impact'][weight]
+                if 'max_impact' in self.config['obo_score_weights'][weight]:
+                    weights[weight] = self.config['obo_score_weights'][weight]['max_impact']
         return weights
 
     def get_description(self):
@@ -271,7 +276,7 @@ def robot_prepare_ontology(o_path, o_out_path, o_metrics_path, base_iris, robot_
     logging.info(f"Preparing {o_path} for dashboard.")
     try:
         callstring = ['robot']
-        callstring.extend(['metrics', robot_opts, '-i', o_path])
+        callstring.extend(['measure', robot_opts, '-i', o_path])
         for prefix in robot_prefixes:
             callstring.extend(['--prefix', f"{prefix}: {robot_prefixes[prefix]}"])
         callstring.extend(['--metrics', 'extended-reasoner','-f','yaml','-o',o_metrics_path, 'merge', 'remove'])
@@ -295,3 +300,99 @@ def count_up(dictionary, value):
 def save_yaml(dictionary, file_path):
     with open(file_path, 'w') as file:
         yaml.dump(dictionary, file)
+
+
+
+def get_hours_since(timestamp):
+    modified_date = datetime.fromtimestamp(timestamp)
+    now = datetime.now()
+    duration = now - modified_date
+    hours_since = (duration.total_seconds() // 3600)
+    return hours_since
+
+def compute_obo_score(impact, reuse, dashboard, weights):
+    impact_weight = weights['impact']
+    reuse_weight = weights['reuse']
+    dashboard_weight = weights['dashboard']
+    sum_weights = impact_weight+reuse_weight+dashboard_weight
+    score_sum = sum([impact_weight*impact, reuse_weight*reuse, dashboard_weight*dashboard])
+    formula = f"({impact_weight}*impact+{dashboard_weight}*dashboard+{reuse_weight}*reuse)/{sum_weights}"
+    score = score_sum/sum_weights
+    return { "score": score, "formula" : formula }
+
+
+def compute_dashboard_score(data, weights, maximpacts):
+
+    if 'failure' in data:
+        return 0
+
+    oboscore = 100
+    no_base = 0
+    report_errors = 0
+    report_warning = 0
+    report_info = 0
+
+    overall_error = 0
+    overall_warning = 0
+    overall_info = 0
+
+    if 'base_generated' in data and data['base_generated'] == True:
+        no_base = weights['base_generated']
+
+    if 'results' in data:
+        if 'ROBOT Report' in data['results']:
+            if 'results' in data['results']['ROBOT Report']:
+                report_errors = data['results']['ROBOT Report']['results']['ERROR']
+                report_warning = data['results']['ROBOT Report']['results']['WARN']
+                report_info = data['results']['ROBOT Report']['results']['INFO']
+
+    if 'summary' in data:
+        overall_error = data['summary']['summary_count']['ERROR']
+        overall_warning = data['summary']['summary_count']['WARN']
+        overall_info = data['summary']['summary_count']['INFO']
+
+    oboscore = oboscore - score_max(weights['no_base'] * no_base, maximpacts['no_base'])
+    oboscore = oboscore - score_max(weights['overall_error'] * overall_error, maximpacts['overall_error'])
+    oboscore = oboscore - score_max(weights['overall_warning'] * overall_warning, maximpacts['overall_warning'])
+    oboscore = oboscore - score_max(weights['overall_info'] * overall_info, maximpacts['overall_info'])
+    oboscore = oboscore - score_max(weights['report_errors'] * report_errors, maximpacts['report_errors'])
+    oboscore = oboscore - score_max(weights['report_warning'] * report_warning, maximpacts['report_warning'])
+    oboscore = oboscore - score_max(weights['report_info'] * report_info, maximpacts['report_info'])
+    return "%.2f" % oboscore
+
+
+def round_float(n):
+    strfloat = "%.2f" % n
+    return (float(strfloat))
+
+def score_max(score,maxscore):
+    if score > maxscore:
+        return maxscore
+    else:
+        return score
+
+def get_prefix_from_url_namespace(ns, curie_map):
+    for prefix in curie_map:
+        if ns==curie_map[prefix]:
+            return prefix
+    raise Exception(f"Namespace {ns} not found in curie map, aborting..")
+
+def get_base_prefixes(curie_map, base_namespaces):
+    internal_ns = []
+    for ns in base_namespaces:
+        prefix = get_prefix_from_url_namespace(ns, curie_map)
+        internal_ns.append(prefix)
+    return internal_ns
+
+def compute_percentage_reused_entities(entity_use_map, internal_ns):
+    internal_entities = 0
+    external_entities = 0
+    for prefix in entity_use_map:
+        if prefix in internal_ns:
+            internal_entities += entity_use_map[prefix]
+        else:
+            external_entities += entity_use_map[prefix]
+
+    reuse_score = 100*(external_entities/(external_entities+internal_entities))
+    score_string = "%.2f" % round(reuse_score, 2)
+    return score_string
